Update Bujaswim.md
Sorry for such a late report, I was adding liquidity to the new contract when this task first appeared, but the guys at github blocked my account by mistake, and all this time I was waiting to restore it. You can see everything by the date of the transactions. Thanks